requirements.txt
# Core NLP & Transformers
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
tokenizers==0.15.0
sentence-transformers==2.2.2

# Traditional NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
python-pptx==0.6.23

# OCR (for scanned documents)
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Table Extraction
camelot-py==0.11.0
tabula-py==2.9.0

# Data Processing & Analysis
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Dashboard
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.66.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Optional: For GPU support (if available)
# torch==2.1.0+cu118  # CUDA 11.8
# Install from: https://pytorch.org/get-started/locally/